refactor: API 요청을 포함한 useCurrentLocation 훅의 위치 변경
