+++ apps/web/src/hooks/api/useCurrentLocation.ts
